+++ R/metacal.R
@@ -2,6 +2,7 @@
 #' 
 #' @name import_phyloseq
 #' @importClassesFrom phyloseq otu_table phyloseq
-#' @importFrom phyloseq otu_table t taxa_are_rows taxa_names sample_names ntaxa
-#'   nsamples sample_sums prune_taxa
+#' @importFrom phyloseq otu_table otu_table<- t taxa_are_rows taxa_names
+#'   sample_names ntaxa nsamples sample_sums prune_taxa 
+#'   sample_data sample_data<-
 NULL